Conflicts:
	gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/widget/refbookmultipicker/client/RefBookMultiPickerView.java

+++ service/src/main/java/com/aplana/sbrf/taxaccounting/service/impl/DepartmentFormTypeServiceImpl.java
@@ -38,6 +38,8 @@ public class DepartmentFormTypeServiceImpl implements DepartmentFormTypeService 
 
     @Override
     public void deleteByIds(List<Long> ids) {
+        if (ids == null || ids.isEmpty())
+            return;
         departmentFormTypeDao.delete(ids);
     }
 }
